add books container component
